1.6dev: use `aiosmtpd` instead of deprecated `smtpd`since Python 3.10 if `aiosmtpd` is available (closes #13510)
git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@17701 af82e41b-90c4-0310-8c96-b1721e28e2e2
