+++ .github/requirements-minimum.txt
@@ -1,4 +1,5 @@
 Jinja2
+aiosmtpd; python_version>='3.10'
 selenium
 pytidylib
 passlib; python_version>='3.13' or python_version<'3.7' or platform_system=='Darwin' or platform_system=='Windows'
